+++ src/scrapers/providers/universal/FardaDownload.js
@@ -1,0 +1,80 @@
+const cheerio = require('cheerio');
+const BaseProvider = require('../BaseProvider');
+
+module.exports = class FardaDownload extends BaseProvider {
+    /** @inheritdoc */
+    getUrls() {
+        return ['https://fardadownload.org'];
+    }
+
+    /** @inheritdoc */
+    async scrape(url, req, ws) {
+        const title = req.query.title.toLowerCase();
+        const year = req.query.year;
+        const season = req.query.season;
+        const episode = req.query.episode;
+        const type = req.query.type;
+        const resolvePromises = [];
+        let headers = {};
+
+        try {
+            const searchTitle = `${title} ${year}`;
+            let searchUrl = (`${url}/?s=${searchTitle.replace(/ /g, '%20')}`);
+            const rp = this._getRequest(req, ws);
+            const jar = rp.jar();
+            const response = await this._createRequest(rp, searchUrl, jar, headers);
+
+            let $ = cheerio.load(response);
+
+            let videoPage = '';
+            $('.post-title a').toArray().forEach(element => {
+
+                let contentTitle = $(element).find('h2').text().toLowerCase();
+                let contentPage = $(element).attr('href');
+
+                if (contentTitle.includes(searchTitle)) {
+                    videoPage = contentPage;
+                }
+            });
+
+            if (!videoPage) {
+                return Promise.resolve();
+            }
+
+            const videoPageHTML = await this._createRequest(rp, videoPage, jar, headers);
+
+            $ = cheerio.load(videoPageHTML);
+
+            if (type == 'tv') {
+                const matchedSection = $(`div.title-season:contains(${season})`).nextUntil('div.title-season', 'div').toArray();
+                for(const section of matchedSection) {
+                    const folderLink = $(section).find('a.downloadf').attr('href');
+                    const folderHTML = await this._createRequest(rp, folderLink, jar, headers);
+                    
+                    $ = cheerio.load(folderHTML);
+
+                    const paddedSeason = `${season}`.padStart(2, '0');
+                    const paddedEpisode = `${episode}`.padStart(2, '0');
+                    const formattedEpisode = `S${paddedSeason}E${paddedEpisode}`;
+
+                    $('a').toArray().forEach(element => {
+                        const fileName = $(element).text();
+                        if (fileName.includes(formattedEpisode)) {
+                            const filePath = $(element).attr('href');
+                            const directLink = this._absoluteUrl(folderLink, filePath);
+                            resolvePromises.push(this.resolveLink(directLink, ws, jar, headers));
+                        }
+                    });
+                }
+            } else {
+                $('a.dl_bx_mv').toArray().forEach(element => {
+                    const directLink = $(element).attr('href');
+                    resolvePromises.push(this.resolveLink(directLink, ws, jar, headers));
+                });
+            }
+        } catch (err) {
+            this._onErrorOccurred(err)
+        }
+        return Promise.all(resolvePromises)
+    }
+}
